refactor: split CLI tests into per-domain test modules
